Complete rewrite; Total API change
I initially wrote this library before I knew what I was doing. Naming
was bad, I was using functions when I should have been using consts, and
I didn't have a good way of handling enharmonic equivalence. I was
testing, but I wasn't using pytest. Essentially I was subverting every
convention. The code was bad, slow, and hard to interface with.

The new design aims to be faster and easier to use. See the README for
usage instructions, which are not exhaustive.

Note that tests have not been added yet, and the old tests were removed.